% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_visualisation.R
\name{data_in_view}
\alias{data_in_view}
\title{data_in_view}
\usage{
data_in_view(my_data, map_coords)
}
\arguments{
\item{my_data}{a dataframe with a \code{lon}(longitude) and \code{lat} 
(latitude) column.}

\item{map_coords}{this should be a list of four numbers, in order: North, 
East, South, West. \code{input$map_bounds} outputs in this order. The
assumption is made that these numbers represent the top left corner and
bottom right corner.}
}
\value{
a new dataframe only containing data within the given coordinates.
}
\description{
subsets the data by given coordinates.
}
\details{
This function can be used to subset tabular data using coordinates 
 from map data. Developed for use with leaflet maps.
}
\author{
Joe Peskett
}
