% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRoot.R
\docType{package}
\name{gRoot-package}
\alias{gRoot}
\alias{gRoot-package}
\title{gRoot: Prototype visualisation for Urban Forest}
\description{
gRoot is a prototype tool for visualising and interacting with 
  the Urban Forest data around Cardiff. Quick start - run `launch_app()`
}
\details{
There are three main components of this package: the data required 
  for the city profile prototype tool, functions for manipulating and
  visualising this data and a shiny app to interact with the data. The
  visualisation functions are used in the shiny app. The functions are
  developed specifically to work with the data for this app and while effort has
  been made to make these functions usable for other datasets, this is not
  completed in all cases.
}
\author{
Joe Peskett
}
\keyword{internal}
