% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{launch_app}
\usage{
launch_app()
}
\value{
launch the gReeneRy app in internet browser window
}
\description{
Visualise and interact with urban forest data of Cardiff
}
\author{
Joe Peskett
}
