% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_hex}
\alias{tree_hex}
\title{A hexmap object for the hexmaps used in city-profile app prototype.}
\format{currently this is in the form of a data frame it has 1498 obs and 51 variables.}
\source{
an amalgamation of NRW estimates of canopy cover in Welsh towns and 
cities, WIMD 2014 data, urban forest data and LSOA 2001 boundaries. The object 
was created using the geogrid package.
}
\usage{
tree_hex
}
\description{
Ideally this will be converted into a JSON eventually so there can be an interactive component.
}
\keyword{datasets}
