% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_visualisation.R
\name{tree_hex_map}
\alias{tree_hex_map}
\title{tree_hex_map}
\usage{
tree_hex_map(data, level, palette_options = "Greens")
}
\arguments{
\item{data}{This MUST be the hex shapefile supplied with 
CityProfileR(default). Other similar files may be created using the 
supplied script.}

\item{level}{What data should be displayed on the hexmap. Currently this 
should be new_col or total_quantile. Can also be any other quantitative
value.}

\item{palette_options}{Any palette that will take quantitative values.}
}
\value{
a hexmap displaying the requested information.
}
\description{
Creates a hexmap for Cardiff - the data for this is stored 
  within a object. The script for making this will be included in the 
  raw_data
}
\details{
Takes data found within the CityProfileR package. Needs to be passed
  new_col or total_quantile
}
\author{
Joe Peskett
}
