% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_visualisation.R
\name{tree_hist}
\alias{tree_hist}
\title{tree_hist}
\usage{
tree_hist(my_data, map_coords, column_name = max_green)
}
\arguments{
\item{my_data}{A dataframe containing tree points. This should have columns 
for max_green, x_coord and y_coord.}

\item{map_coords}{A list containing 4 coordinates, representing the top left 
and bottom right points of a bounding box.}
}
\value{
A histogram of data that lies within the defined bounding box.
}
\description{
Creates a histogram of data that falls within a given bounding 
  box
}
\details{
For this function it is important that columns of the data frame are
  correctly labeled "x_coord" and "y_coord" and that in map_coords the order
  of coordinates is correct. This is used to subset my_data
  
  my_data is then placed as the data argument in ggplot2. \$max_green is 
  selected for the aesthetic and style arguments are given to geom_histogram.
  
  Note: tree_hist currently throws a small error in the console - this is 
  because there is a log1- transformation causing some non-finite values.
}
\author{
Joe Peskett
}
