% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_visualisation.R
\name{tree_plot}
\alias{tree_plot}
\title{tree_plot}
\usage{
tree_plot(my_data = tree_points, column_name = road, column_filter,
  plotting_vars = c("left_green", "right_green"), lon_val = "lon",
  lat_val = "lat")
}
\arguments{
\item{my_data}{The dataframe you would like to plot. Default is tree_points.}

\item{column_name}{The name of column you want to subset the data by. Default
is road.}

\item{column_filter}{The value of the selected column to filter the data by.}

\item{plotting_vars}{This should be a character vector of length 2. These are
the variables that will be plotted above and below. The first value will be
plotted above zero, the second will be plotted below. They should be in
approximately in the same scale. Default is left_green and right_green.}

\item{lon_val}{the value for longitude. Default is lon.}

\item{lat_val}{the value for lattitude. Default is lat.}
}
\value{
A ggplot graphic showing left side of the road above the line, right 
  side on the bottom of the graph.
}
\description{
Area plot to show amount of variable on two sides of a road.
}
\details{
Data is manipulated using dplyr::mutate, adds a new column by 
  multiplying the second variable by 1. Then, subsets by the 
  \code{column_name} that has been selected in \code{column_filter}. The 
  final bit of manipulation uses melt from the reshape2 package to transform 
  dataframe into long format which is required for ggplot2.
  
  The end product of this is used as the data for ggplot2. Add the 
  aesthetics, geoms and labels required for the plot.
  
  This function was created for use with the city profile prototype however 
  has been adjusted so that other data sets with alternative column names may
  also be chosen.
  
  Note that currently the x axis used is the absolute value of longitude * 
  lattitude. This is an quick and ultimately flawed attempt to deal with the 
  fact with the fact that roads are rarely completely straight or all facing 
  the same direction.
}
\author{
Joe Peskett
}
