% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_visualisation.R
\name{tree_plotly}
\alias{tree_plotly}
\title{tree_plotly}
\usage{
tree_plotly(my_data = tree_points, column_name = road, column_filter)
}
\description{
DNA plot of selected road, broken down by way_id.
}
\author{
Joe Peskett
}
