% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_points}
\alias{tree_points}
\title{A dataset containing 110034 points around the road network of Cardiff}
\format{A data frame with 110034 rows and 9 variables:
\describe{
    \item{left_green}{Percentage vegetation on left side of road}
    \item{right_green}{Percentage vegetation on right side of road}
    \item{road}{Name of the road}
    \item{way_id}{Unique id of road}
    \item{seq}{Sequence of recordings along the specific road}
    \item{max_green}{Value of the highest amount of green (left or right) for the point}
    \item{parish}{Parish this point fall in}
    \item{lon}{Longitude of that point}
    \item{lat}{Latitude of that point}}}
\source{
Data will be available on github after it has been published.
}
\usage{
tree_points
}
\description{
A dataset containing 110034 points around the road network of Cardiff
}
\keyword{datasets}
