% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_polygons}
\alias{tree_polygons}
\title{Files for displaying polygons in leaflet map}
\format{A SpatialPolygonsDataFrame with the shapes of Cardiff Parishes and they're mean vegetation value}
\source{
Combination of Parish boundaries from ONS geohub combined with urban forests data. Created using functions found in CityProfileR.
}
\usage{
tree_polygons
}
\description{
Files for displaying polygons in leaflet map
}
\keyword{datasets}
