% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_visualisation.R
\name{tree_road_shape}
\alias{tree_road_shape}
\title{tree_road_shape}
\usage{
tree_road_shape(data = tree_points, column_name = road, column_filter,
  colour_adjust = way_id, size_adjust = max_green)
}
\arguments{
\item{data}{Dateframe containing geospatial data. Should have columns 
containing longitude and lattitude data. Default is tree_points.}

\item{column_name}{the column that you want to filter data by. Default is
road.}

\item{column_filter}{a value from the specified column that you want to keep.}

\item{colour_adjust}{a variable in the dataset that you want to vary the
colours of points by. Default is max_green.}

\item{size_adjust}{a variable in the dataset that you want to vary the size
of points by. Default is max_green.}
}
\value{
a ggplot with the shape of the selected street.
}
\description{
Creates a plot where size and colour are based on the max_green 
  value. X and Y are taken as the lon and lat values.
}
\details{
Data is found within the package CityProfileR. Can take the argument
  to specify a street.
}
\author{
Joe Peskett
}
