% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_visualisation.R
\name{tree_table}
\alias{tree_table}
\title{tree_table}
\usage{
tree_table(my_data = tree_points, col_name, map_coords,
  col_1 = max_green, col_2 = left_green, col_3 = right_green)
}
\arguments{
\item{my_data}{A dataframe the full data set to be displayed. Default is 
tree_points.}

\item{col_name}{The name of the column to group by. Usually this will be 
"road" or "parish"}

\item{map_coords}{this should be a list of four numbers, in order: North, 
East, South, West. \code{input$map_bounds} outputs in this order. The 
assumption is made that these numbers represent the top left corner and 
bottom right corner.}

\item{col_1}{a column from the data set to be summarised. Default is max_green.}

\item{col_2}{a column from the data set to be summarised. Default is left_green.}

\item{col_3}{a column from the data set to be summarised. Default is right_green.}
}
\value{
A new dataframe that can then be rendered in a shiny app. Columns
  values will just be named 1-3.
}
\description{
Creates a table that displays the data falls within a given 
  bounding box
}
\details{
This function first subsets 'my_data' by comparing 'x_coords' and 
  'y_coords'with 'map_coords'. It is important that  columns of the data 
  frame are correctly labeled "x_coord" and "y_coord" and that in map_coords 
  the order of coordinates is correct. This should be in a list.
  
  The data is then grouped by \code{col_name}. A summary of this data is then
  create to show number of records, mean left_green, mean right_green and 
  mean max_green. Note: It is important to note that these summaries will be 
  for whatever is currently in view of the map. For particularly long roads, 
  the summary statistics will be for stretch of road that is in view of the 
  map.
}
\author{
Joe Peskett
}
